% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvanceRConnectingToAPI.R
\docType{class}
\name{AdvanceRConnectingToAPI-class}
\alias{AdvanceRConnectingToAPI-class}
\alias{AdvanceRConnectingToAPI}
\title{A Reference Class to pull data for Sweden economy from Kolada.}
\description{
This class extracts KPI data from kolada across multiple municipality and/or years.
}
\section{Fields}{

\describe{
\item{\code{kpi}}{List. Captures the list of codes of one or more KPI for which the data is to be extracted.}

\item{\code{year}}{List. Captures the list of one or more year for which the data is to be extracted.}

\item{\code{municipality}}{List. Captures the list of codes of one or more municipality for which the data is to be extracted.}

\item{\code{searchText}}{Character. String to be used as search string when identifying KPI(s) or municipality(ies).}

\item{\code{searchTextIsKPI}}{Logical. (Default=TRUE). FALSE=Municipality Search, TRUE=KPI search. Used only when searchText is specified.}

\item{\code{sweText}}{Logical. (Default=TRUE). TRUE when search string is provided in Swedish language but FALSE when provided in English (takes longer to process). Available only for KPIs. Note: Translation is done using "mymemory API". Respective restrictions apply.}

\item{\code{url}}{Character. The URL that was used to fetch the data.}

\item{\code{fetchedData}}{Data.frame The fetched data stored in a structred data frame.}

\item{\code{status}}{Numeric. The http request response code.}
}}


\examples{
\dontrun{
  apiCallKolada <- AdvanceRConnectingToAPI$new(kpi = list("N25026"), year = list("2020", "2022"))
  apiCallKolada$summaryStatsData()
}

}
